Added compressor audio mixer group. Added mixer group to lamp control.
