+++ CCL.Types/Proxies/Indicators/LampControlProxy.cs
@@ -1,18 +1,16 @@
-﻿using UnityEngine;
+﻿using CCL.Types.Proxies.Audio;
+using UnityEngine;
 
 namespace CCL.Types.Proxies.Indicators
 {
     public class LampControlProxy : MonoBehaviour
     {
-        public IndicatorEmissionProxy lampInd;
-
+        public IndicatorEmissionProxy lampInd = null!;
         public LampState lampState;
-
-        public AudioClip warningAudio;
-
-        public AudioClip onStateBuzzingLoopAudio;
-
-        public AudioClip blinkStateBuzzingLoopAudio;
+        public AudioClip warningAudio = null!;
+        public AudioClip onStateBuzzingLoopAudio = null!;
+        public AudioClip blinkStateBuzzingLoopAudio = null!;
+        public DVAudioMixerGroup audioMixerGroup = DVAudioMixerGroup.Cab;
 
         public enum LampState
         {
